fixing folders part 2
