+++ Dayana/Server/Api/Extensions/DependencyInjection/ServiceInjection.cs
@@ -1,5 +1,5 @@
-﻿using Dayana.Shared.Persistence;
-using Dayana.Shared.Persistence.Interfaces;
+﻿using Dayana.Shared.Persistence.EntityFrameWorkObjects.RepositoryObjects.Interfaces.UnitOfWorks;
+using Dayana.Shared.Persistence.EntityFrameWorkObjects.RepositoryObjects.Repositories.UnitOfWorks;
 
 namespace Dayana.Server.Api.Extensions.DependencyInjection;
 
@@ -7,7 +7,7 @@ public static class ServiceInjection
 {
     public static IServiceCollection AddServices(this IServiceCollection services)
     {
-        services.AddScoped<IUnitOfWork, UnitOfWork>();
+        services.AddScoped<IUnitOfWorkIdentity, UnitOfWorkIdentity>();
 
         return services;
     }
