+++ Dayana/Shared/Persistence/EntityFrameWorkObjects/RepositoryObjects/Interfaces/IdentityRepositories/IClaimRepository.cs
@@ -1,7 +1,7 @@
 ﻿using Dayana.Shared.Domains.Identity.Claims;
 using Dayana.Shared.Persistence.Models.Filters.Claims;
 
-namespace Dayana.Shared.Persistence.Interfaces.IdentityRepositories;
+namespace Dayana.Shared.Persistence.EntityFrameWorkObjects.RepositoryObjects.Interfaces.IdentityRepositories;
 
 public interface IClaimRepository : IRepository<Claim>
 {
