+++ Dayana/Shared/Persistence/EntityFrameWorkObjects/RepositoryObjects/Interfaces/IdentityRepositories/IPermissionRepository.cs
@@ -1,7 +1,7 @@
 ﻿using Dayana.Shared.Domains.Identity.Permissions;
 using Dayana.Shared.Persistence.Models.Filters.Permissions;
 
-namespace Dayana.Shared.Persistence.Interfaces.IdentityRepositories;
+namespace Dayana.Shared.Persistence.EntityFrameWorkObjects.RepositoryObjects.Interfaces.IdentityRepositories;
 
 public interface IPermissionRepository : IRepository<Permission>
 {
