+++ Dayana/Shared/Persistence/EntityFrameWorkObjects/Seeding/IdentitySeeds/RolePermissionSeed.cs
@@ -1,6 +1,6 @@
 ﻿using Dayana.Shared.Domains.Identity.Roles;
 
-namespace Dayana.Shared.Persistence.Seeding.IdentitySeeds;
+namespace Dayana.Shared.Persistence.EntityFrameWorkObjects.Seeding.IdentitySeeds;
 
 public static class RolePermissionSeed
 {
