+++ Dayana/Shared/Persistence/EntityFrameWorkObjects/Seeding/IdentitySeeds/UserRoleSeed.cs
@@ -1,6 +1,6 @@
 ﻿using Dayana.Shared.Domains.Identity.Users;
 
-namespace Dayana.Shared.Persistence.Seeding.IdentitySeeds;
+namespace Dayana.Shared.Persistence.EntityFrameWorkObjects.Seeding.IdentitySeeds;
 
 public static class UserRoleSeed
 {
